LibC: Make it compile on 64-bit
It won't actually work on 64-bit yet, but let's do our future selves
a favor and make it compile.

+++ Userland/Libraries/LibC/mman.cpp
@@ -35,7 +35,7 @@ extern "C" {
 void* serenity_mmap(void* addr, size_t size, int prot, int flags, int fd, off_t offset, size_t alignment, const char* name)
 {
     Syscall::SC_mmap_params params { (uintptr_t)addr, size, alignment, prot, flags, fd, offset, { name, name ? strlen(name) : 0 } };
-    ssize_t rc = syscall(SC_mmap, &params);
+    ptrdiff_t rc = syscall(SC_mmap, &params);
     if (rc < 0 && -rc < EMAXERRNO) {
         errno = -rc;
         return MAP_FAILED;
@@ -56,7 +56,7 @@ void* mmap_with_name(void* addr, size_t size, int prot, int flags, int fd, off_t
 void* mremap(void* old_address, size_t old_size, size_t new_size, int flags)
 {
     Syscall::SC_mremap_params params { (uintptr_t)old_address, old_size, new_size, flags };
-    ssize_t rc = syscall(SC_mremap, &params);
+    ptrdiff_t rc = syscall(SC_mremap, &params);
     if (rc < 0 && -rc < EMAXERRNO) {
         errno = -rc;
         return MAP_FAILED;
@@ -95,7 +95,7 @@ int madvise(void* address, size_t size, int advice)
 
 void* allocate_tls(size_t size)
 {
-    int rc = syscall(SC_allocate_tls, size);
+    ptrdiff_t rc = syscall(SC_allocate_tls, size);
     if (rc < 0 && -rc < EMAXERRNO) {
         errno = -rc;
         return MAP_FAILED;
